// examples/boilerplate/src/types.ts
export type ModeType = 'light' | 'dark' | 'system';

export type OrientationType = 'vertical' | 'horizontal';

export type DirectionType = 'ltr' | 'rtl';

export type LocaleType = (typeof i18n)['locales'][number];

export type ThemeType = keyof typeof themes;